feat: update nesting layout page
