+++ src/app/examples/nextjs/routing/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 export default function Page() {
   return <div>Next.js Routing</div>;
 }
